+++ integration/pom.xml
@@ -18,10 +18,12 @@
         <dependency>
             <artifactId>web</artifactId>
             <groupId>fit-goal</groupId>
+            <scope>test</scope>
         </dependency>
         <dependency>
             <groupId>io.dropwizard</groupId>
             <artifactId>dropwizard-testing</artifactId>
+            <scope>test</scope>
         </dependency>
         <dependency>
             <groupId>org.junit.jupiter</groupId>
